Add support for XF86Battery button (Bug #14055)

+++ src/xfpm-button.h
@@ -53,7 +53,7 @@ GType                 		xfpm_button_get_type               	(void) G_GNUC_CONST;
 
 XfpmButton       	       *xfpm_button_new             		(void);
 
-guint8                		xfpm_button_get_mapped		     	(XfpmButton *button) G_GNUC_PURE;
+guint16               		xfpm_button_get_mapped		     	(XfpmButton *button) G_GNUC_PURE;
 
 G_END_DECLS
